Release 1.0
Excellent news! I'm finally done with the code for this project. Cleaned up some files, as well. Now I just have to write the report....

Also, I might try to further tweak the neural network in my free time, cuz this is kinda fun in a weird way. If I do, I promise I'll do it in a separate file.